chore(Accordion): rename AccordionProvider files to AccordionGroup
